requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-core
langchain_community
langchain-google-genai
google-generativeai
python-dotenv
faiss-cpu
sentence-transformers
pydantic
python-multipart
requests
numpy
gunicorn
